requirements.txt
pytest~=8.2.2
selenium~=4.21.0
openpyxl~=3.0.0
pytest-html~=4.1.1
pytest-xdist~=2.2.1
allure-pytest~=2.13.5